Enum linera_alloy_json_rpc::RpcError

pub enum RpcError<E, ErrResp = Box<RawValue>> {
    ErrorResp(ErrorPayload<ErrResp>),
    NullResp,
    UnsupportedFeature(&'static str),
    LocalUsageError(Box<dyn Error + Send + Sync>),
    SerError(Error),
    DeserError {
        err: Error,
        text: String,
    },
    Transport(E),
}

An RPC error.

Variants

ErrorResp(ErrorPayload<ErrResp>)

Server returned an error response.

NullResp

Server returned a null response when a non-null response was expected.

UnsupportedFeature(&'static str)

Rpc server returned an unsupported feature.

LocalUsageError(Box<dyn Error + Send + Sync>)

Returned when a local pre-processing step fails. This allows custom errors from local signers or request pre-processors.

SerError(Error)

JSON serialization error.

Tuple Fields

0: Error

The underlying serde_json error.

DeserError

JSON deserialization error.

Fields

err: Error

The underlying serde_json error.

text: String

For deser errors, the text that failed to deserialize.

Transport(E)

Transport error.

This variant is used when the error occurs during communication.

Tuple Fields

0: E

The underlying transport error.

Implementations

impl<E, ErrResp> RpcError<E, ErrResp>
where ErrResp: RpcReturn,

pub const fn err_resp(err: ErrorPayload<ErrResp>) -> Self

Instantiate a new ErrorResp from an error response.

pub fn local_usage(err: impl Error + Send + Sync + 'static) -> Self

Instantiate a new LocalUsageError from a custom error.

pub fn local_usage_str(err: &str) -> Self

Instantiate a new LocalUsageError from a custom error message.

pub fn deser_err(err: Error, text: impl AsRef<str>) -> Self

Instantiate a new DeserError from a serde_json::Error and the text. This should be called when the error occurs during deserialization.

Note: This will check if the response is actually an ErrorPayload, if so it will return a RpcError::ErrorResp.

impl<E, ErrResp> RpcError<E, ErrResp>

pub const fn ser_err(err: Error) -> Self

Instantiate a new SerError from a serde_json::Error. This should be called when the error occurs during serialization.

pub const fn is_ser_error(&self) -> bool

Check if the error is a serialization error.

pub const fn is_deser_error(&self) -> bool

Check if the error is a deserialization error.

pub const fn is_transport_error(&self) -> bool

Check if the error is a transport error.

pub const fn is_error_resp(&self) -> bool

Check if the error is an error response.

pub const fn is_null_resp(&self) -> bool

Check if the error is a null response.

pub const fn is_unsupported_feature(&self) -> bool

Check if the error is an unsupported feature error.

pub const fn is_local_usage_error(&self) -> bool

Check if the error is a local usage error.

pub const fn as_error_resp(&self) -> Option<&ErrorPayload<ErrResp>>

Fallible conversion to an error response.