Struct linera_alloy_json_rpc::Request

pub struct Request<Params> {
    pub meta: RequestMeta,
    pub params: Params,
}

A JSON-RPC 2.0 request object.

This is a generic type that can be used to represent any JSON-RPC request. The Params type parameter is used to represent the parameters of the request, and the method field is used to represent the method name.

Note

The value of method should be known at compile time.

Fields
meta: RequestMeta

The request metadata (ID and method).

params: Params

The request parameters.

Implementations

impl<Params> Request<Params>

pub fn new(method: impl Into<Cow<'static, str>>, id: Id, params: Params) -> Self

Create a new Request.

pub fn is_subscription(&self) -> bool

Returns true if the request is a subscription.

pub fn set_is_subscription(&mut self)

Indicates that the request is a non-standard subscription (i.e. not โeth_subscribeโ).

pub fn set_subscription_status(&mut self, sub: bool)

Setter for is_subscription. Indicates to RPC clients that the request triggers a stream of notifications.

impl<Params> Request<Params>
where Params: RpcParam,

pub fn box_params(self) -> PartiallySerializedRequest

Serialize the request parameters as a boxed RawValue.

Panics

If serialization of the params fails.

pub fn serialize(self) -> Result<SerializedRequest>

Serialize the request, including the request parameters.

impl<Params> Request<&Params>
where Params: Clone,

pub fn into_owned_params(self) -> Request<Params>

Clone the request, including the request parameters.

impl<'a, Params> Request<Params>
where Params: AsRef<RawValue> + 'a,

pub fn try_params_as<T: DeserializeOwned>(&self) -> Result<T>

Attempt to deserialize the params.

To borrow from the params via the deserializer, use Request::try_borrow_params_as.

Returns

Ok(T) if the params can be deserialized as T

Err(e) if the params cannot be deserialized as T

pub fn try_borrow_params_as<T: Deserialize<'a>>(&'a self) -> Result<T>

Attempt to deserialize the params, borrowing from the params

Returns

Ok(T) if the params can be deserialized as T

Err(e) if the params cannot be deserialized as T