Type Alias linera_alloy_json_rpc::BorrowedResponsePayload

pub type BorrowedResponsePayload<'a> = ResponsePayload<&'a RawValue, &'a RawValue>;

A ResponsePayload that has been partially deserialized, borrowing its contents from the deserializer. This is used primarily for intermediate deserialization. Most users will not require it.

See the top-level docs for more info.

Aliased Type

enum BorrowedResponsePayload<'a> {
    Success(&'a RawValue),
    Failure(ErrorPayload<&'a RawValue>),
}

Variants

Success(&'a RawValue)

A successful response payload.

Failure(ErrorPayload<&'a RawValue>)

An error response payload.

Implementations

impl BorrowedResponsePayload<'_>

pub fn into_owned(self) -> ResponsePayload

Convert this borrowed response payload into an owned payload by copying the data from the deserializer (if necessary).